Add MVP Frontend Skeleton
This frontend is very ugly, but it works as an MVP. All it does
is, on init, query all for all the entries in the PlayerStatistics
database via the API exposed by our backend. It then displays
these statistics (in a currently quite... unaesthetic way) as
cards in the home page.